[grapheneos conflict resolution] ep_create_wakeup_source(): dentry name can change under you...
fix conflict with 875d9da511554bb2457cb7b177a13f5b4cf3d33e and other
upstream merges
also don't change the arguments passed to wakeup_source_register as per upstream commit as we pass an argument to it (buf) that is different from upstream (upstream passes name directly and needs it to be changed to n.name). we get buf by passing name into the snprintf call above. simply change the argument in that snprintf call to n.name

commit 3701cb59d892b88d569427586f01491552f377b1 upstream.

or get freed, for that matter, if it's a long (separately stored)
name.

Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: anupritaisno1 <www.anuprita804@gmail.com>